+++ server/src/main/java/com/dynabyte/marleyrest/registration/RegistrationUseCase.java
@@ -2,16 +2,16 @@ package com.dynabyte.marleyrest.registration;
 
 import com.dynabyte.marleyrest.api.util.RequestUtil;
 import com.dynabyte.marleyrest.prediction.PredictionUseCase;
-import com.dynabyte.marleyrest.prediction.request.ImageRequest;
-import com.dynabyte.marleyrest.prediction.response.FaceRecognitionResponse;
-import com.dynabyte.marleyrest.prediction.service.FaceRecognitionService;
+import com.dynabyte.marleyrest.personrecognition.request.ImageRequest;
+import com.dynabyte.marleyrest.personrecognition.response.FaceRecognitionResponse;
+import com.dynabyte.marleyrest.personrecognition.service.FaceRecognitionService;
 import com.dynabyte.marleyrest.registration.exception.MissingPersonInDbException;
 import com.dynabyte.marleyrest.registration.exception.PersonAlreadyInDbException;
 import com.dynabyte.marleyrest.registration.exception.RegistrationException;
-import com.dynabyte.marleyrest.registration.model.Person;
+import com.dynabyte.marleyrest.personrecognition.model.Person;
 import com.dynabyte.marleyrest.registration.request.LabelPutRequest;
 import com.dynabyte.marleyrest.registration.request.RegistrationRequest;
-import com.dynabyte.marleyrest.registration.service.PersonService;
+import com.dynabyte.marleyrest.personrecognition.service.PersonService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
